requirements.txt
Flask==0.10.1
Jinja2==2.7.1
MarkupSafe==0.18
Werkzeug==0.9.3
argparse==1.2.1
distribute==0.6.24
itsdangerous==0.23
pycrypto==2.6
wsgiref==0.1.2
